train/requirements.txt
boto3==1.18.65
botocore==1.21.65
importlib-metadata==4.8.1
loguru==0.5.3
lxml==4.6.5
mock~=4.0.3
mypy-extensions==0.4.3
packaging==21.3
pathspec==0.9.0
platformdirs==2.4.0
pandas==1.4.1
pg8000==1.23.0
py==1.10.0
pyarrow==6.0.1
pytest==6.2.5
python-dotenv==0.19.0
PyYAML==5.4.1
redshift-connector==2.0.901
requests==2.26.0
setuptools~=28.8.0
s3transfer==0.5.0
tqdm==4.62.3
joblib==1.0.1
torch==1.12.0
transformers==4.21.0
scikit-learn==1.0.2
